Split count_distinct.rs into separate modules
